feat: docker version
