--- bot/database/mysql_db.py
@@ -1,1 +1,0 @@
-#soon
